changed maps to be a separate component that renders on each spot page but still getting console errors even though functional
