+++ frontend/src/components/ViewOneSpot/index.js
@@ -12,6 +12,7 @@ import Col from 'react-bootstrap/Col';
 import Row from 'react-bootstrap/Row';
 import './ViewOneSpot.css';
 import { GoogleMap, LoadScript, Marker } from '@react-google-maps/api';
+import MyMaps from '../MyMaps';
 
 
 
@@ -40,60 +41,6 @@ const ViewOneSpot = () => {
         history.push('/')
     }
 
-    const mapStyles = {
-        height: "40vh",
-        width: "50%"
-    };
-
-
-    const locations = [
-        {
-          id: 1,
-          location: {
-            lat: 59.3409,
-            lng: 18.0754
-          },
-        },
-        {
-          id: 2,
-          location: {
-            lat: 59.3313,
-            lng: 18.0311
-          },
-        },
-        {
-          id: 3,
-          location: {
-            lat: 19.734005,
-            lng: -155.040745
-          },
-        },
-        {
-          id: 4,
-          location: {
-            lat: 35.90016,
-            lng: 14.51000
-          },
-        },
-        {
-          id: 5,
-          location: {
-            lat: 36.0312,
-            lng: 14.2180
-          },
-        },
-        {
-            id: 6,
-            location: {
-              lat: 41.4055,
-              lng: 2.1915
-            },
-        }
-    ];
-
-    const newCenter = locations.find(item => item.id === id).location
-    // console.log(newCenter)
-
 
     return (
         <div>
@@ -113,26 +60,18 @@ const ViewOneSpot = () => {
                                 </Col>
                             )
                         })}
-                        <GoogleMap googleMapsApiKey='AIzaSyAUN3kd3mSSvon94EFE-xpunp5dWR20gUM'
+                        <MyMaps props={{spotState}}/>
+                        {/* <GoogleMap googleMapsApiKey='AIzaSyAUN3kd3mSSvon94EFE-xpunp5dWR20gUM'
                             mapContainerStyle={mapStyles}
                             zoom={16}
                             center={newCenter}
 
                             >
                             <Marker position={newCenter}/>
-                        </GoogleMap>
+                        </GoogleMap> */}
                     </Row>
-
                 </Container>
             </div>
-            {/* <GoogleMap googleMapsApiKey='AIzaSyAUN3kd3mSSvon94EFE-xpunp5dWR20gUM'
-                mapContainerStyle={mapStyles}
-                zoom={15}
-                center={newCenter}
-
-            >
-                <Marker position={newCenter}/>
-            </GoogleMap> */}
             <div className='spotPageContainer'>
                 <h4>{spotState.location}</h4>
                 <h4>${spotState.price}</h4>
@@ -158,5 +97,3 @@ const ViewOneSpot = () => {
 
 
 export default ViewOneSpot;
-
-// Google Maps apiKey: 'AIzaSyAUN3kd3mSSvon94EFE-xpunp5dWR20gUM'
